fix(middleware): address Codex review — cache entry cap + TTL type
Two regressions flagged on PR #114 by Codex:

P1 (cache entry cap lost): built-in ResponseCachingMiddleware defaults
to an unbounded in-memory store, dropping the `response_cache_max`
guardrail our custom middleware used to enforce. Under high-cardinality
tool arguments the cache could grow unboundedly. Fix: pass
`cache_storage=MemoryStore(max_entries_per_collection=
settings.mcp.response_cache_max)` — py-key-value-aio's MemoryStore
supports a per-collection LRU cap, preserving the original behaviour.
For distributed deployments, swap the storage backend for Redis or
FileTree.

P2 (TTL truncation): `ttl=int(settings.mcp.response_cache_ttl)` silently
truncated fractional floats (0.5 → 0), which would have disabled
caching entirely for those configurations. CallToolSettings.ttl is a
TypedDict[int] slot. Fix: change MCPSettings.response_cache_ttl to
`int` at source (default 60, range 0..3600) so no cast is needed. The
legacy `response_cache_ttl_s: int` alias already existed — this aligns
them on the same type.

Sub-second TTL is not a meaningful use case for our tool-call cache;
going int-only is both safer and simpler.

+++ app/config/mcp.py
@@ -19,8 +19,11 @@ class MCPSettings(BaseSettings):
     debug: bool = Field(default=False, description="Surface raw exception details to clients.")
     pagination_size: int = Field(default=100, ge=10, le=1000)
     response_cache_ttl_s: int = Field(default=60, ge=0, le=3600)
-    # Aliases aligned with Phase 5 plan naming (float TTL + LRU cap).
-    response_cache_ttl: float = Field(default=60.0, ge=0.0, le=3600.0)
+    # FastMCP's built-in ``ResponseCachingMiddleware`` expects integer TTL
+    # seconds (``CallToolSettings.ttl: int``). Keeping this field an int avoids
+    # silent truncation from float values like 0.5 → 0 (which would disable
+    # caching). Sub-second TTL is not a real use case for our tool-call cache.
+    response_cache_ttl: int = Field(default=60, ge=0, le=3600)
     response_cache_max: int = Field(default=1024, ge=1, le=100_000)
     response_size_limit_bytes: int = Field(default=400_000, ge=1000)
     response_max_bytes: int = Field(default=400_000, ge=1000)
